Problem Statement-2/requirements.txt
openai>=1.0
websockets>=12.0
sounddevice>=0.4
numpy>=1.24
python-dotenv>=1.0
